Rename route guards
