+++ bitwarden_license/src/app/organizations/organizations-routing.module.ts
@@ -6,8 +6,8 @@ import { Permissions } from "jslib-common/enums/permissions";
 
 import { OrganizationLayoutComponent } from "src/app/layouts/organization-layout.component";
 import { ManageComponent } from "src/app/organizations/manage/manage.component";
-import { OrganizationRedirectGuardService } from "src/app/services/organization-guard.service";
-import { OrganizationPermissionsGuardService } from "src/app/services/organization-type-guard.service";
+import { OrganizationPermissionsGuardService } from "src/app/services/organization-permissions-guard.service";
+import { OrganizationRedirectGuardService } from "src/app/services/organization-redirect-guard.service";
 
 import { SsoComponent } from "./manage/sso.component";
